requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
APScheduler==3.8.0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
ciso8601==2.2.0
click==8.0.1
colorlog==6.4.1
docopt==0.6.2
ed25519==1.5
git+https://github.com/hikari-py/hikari
hikari-lightbulb==1.4.0
idna==3.2
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
pipreqs==0.4.10
platformdirs==2.4.0
pure25519==0.0.1
pycares==4.0.0
pycparser==2.20
pytz==2021.1
regex==2021.9.24
requests==2.26.0
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
uvloop==0.16.0
yarg==0.1.9
yarl==1.6.3
